+++ lib/kanji_dictionary.dart
@@ -5,7 +5,7 @@ export 'src/kanji_dictionary.dart';
 export 'src/language.dart';
 export 'src/character.dart';
 export 'src/difficulty.dart';
-export 'src/meaning.dart';
+export 'src/meanings.dart';
 export 'src/reading.dart';
 export 'src/readings.dart';
 export 'src/book_index.dart';