+++ lib/src/character.dart
@@ -3,46 +3,39 @@ import 'package:xml/xml.dart';
 import 'book_index.dart';
 import 'difficulty.dart';
 import 'language.dart';
-import 'meaning.dart';
+import 'meanings.dart';
 import 'reading.dart';
 import 'readings.dart';
 
 class Character {
   final String literal;
   final Difficulty difficulty;
-  final List<Meaning> meanings;
+  final Meanings _meanings;
   final Readings _readings;
   final BookIndex index;
 
   Character(
       {required this.literal,
       required this.difficulty,
-      required this.meanings,
-      required readings,
+      required Meanings meanings,
+      required Readings readings,
       required this.index})
-      : _readings = readings;
+      : _meanings = meanings,
+        _readings = readings;
 
   List<String> getMeanings(Language language) {
-    return meanings
-        .where((m) => m.language == language)
-        .map((m) => m.meaning)
-        .toList();
+    return _meanings.meanings[language] ?? [];
   }
+
+  Map<Language, List<String>> get meanings => _meanings.meanings;
 
   Map<Reading, List<String>> get readings => _readings.readings;
 
   factory Character.fromXml(XmlElement el) {
-    final readingMeaning = el.getElement('reading_meaning');
     return Character(
         literal: el.getElement('literal')!.text,
         difficulty: Difficulty.fromXml(el.getElement('misc')!),
-        meanings: readingMeaning != null
-            ? readingMeaning
-                .getElement('rmgroup')!
-                .findElements('meaning')
-                .map((e) => Meaning.fromXml(e))
-                .toList()
-            : [],
+        meanings: Meanings.fromXml(el),
         readings: Readings.fromXml(el),
         index: BookIndex.fromXml(el.getElement('dic_number')));
   }